backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-socketio==5.11.4
anthropic==0.42.0
httpx==0.28.1
asyncpg==0.30.0
python-dotenv==1.0.1
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.159.0
serpapi==0.1.5
deepgram-sdk==3.8.0
beautifulsoup4==4.12.3
pydantic==2.10.4
rembg==2.0.57
Pillow==11.0.0
pytest==8.3.4
pytest-asyncio==0.24.0
